Sort library link *second* when tabs == true
When the library link is inserted while the (Material)
site uses tabs the link has to be inserted in *second*
position because the first element will used (by MkDocs)
both as the first element in the navigation and the link
for the tab element. So this has to be the first sub-page.

This is not ideal since the library link is injected after the
first actual page. But since that first page is usually the
home page of the section, I think it's OK.