Parse template as YAML
Closes #6
Closes #7

This commit first replaces the template fields (as before)
but then loads the template as a YAML dictionary, which
is then later serialized for writing the mkdocs.yml files.
Therefore it is now possible to access the configuration
not only by self.config() anymore (which required any
interesting stuff to be explicitly written to the book-config.yml
file) but through the self.common() dict.